+++ frontend/search_cms/lib/features/dashboard/data/data_sources/insert_artifact_api_impl.dart
@@ -38,8 +38,8 @@ class InsertArtifactApiImpl implements AbstractInsertArtifactApi {
   Future<void> insertArtifact({
     required String assemblageName,
     int? porosity,
-    int? sizeUpper,
-    int? sizeLower,
+    double? sizeUpper,
+    double? sizeLower,
     String? comment,
     int? preExcavFrags,
     int? postExcavFrags,
