+++ frontend/search_cms/lib/features/dashboard/data/repositories/insert_artifact_repository_impl.dart
@@ -41,8 +41,8 @@ class InsertArtifactRepositoryImpl implements AbstractInsertArtifactRepository {
   Future<Result> insertArtifact({
     required String assemblageName,
     int? porosity,
-    int? sizeUpper,
-    int? sizeLower,
+    double? sizeUpper,
+    double? sizeLower,
     String? comment,
     int? preExcavFrags,
     int? postExcavFrags,
