+++ frontend/search_cms/lib/features/dashboard/domain/repositories/abstract_insert_artifact_repository.dart
@@ -27,8 +27,8 @@ abstract class AbstractInsertArtifactRepository {
   Future<Result> insertArtifact({
     required String assemblageName,
     int? porosity,
-    int? sizeUpper,
-    int? sizeLower,
+    double? sizeUpper,
+    double? sizeLower,
     String? comment,
     int? preExcavFrags,
     int? postExcavFrags,
