+++ frontend/search_cms/lib/features/dashboard/domain/usecases/insert_artifact_usecase.dart
@@ -39,8 +39,8 @@ class InsertArtifactUsecase {
   Future<Result> call({
     required String assemblageName,
     int? porosity,
-    int? sizeUpper,
-    int? sizeLower,
+    double? sizeUpper,
+    double? sizeLower,
     String? comment,
     int? preExcavFrags,
     int? postExcavFrags,
